refactor: dissolve utils package into specialized packages
The utils package was a grab-bag of unrelated helpers. Split it into
cohesive packages for multi-consumer helpers and move single-consumer
helpers into their one consumer, folding in Go 1.26 stdlib cleanups.

New packages (under internal/, named to avoid stdlib collisions):
- internal/fsutil: FileExists (now errors.Is), MustGetwd, MustGetUserHomeDir
- internal/envutil: ConvertEnv, ConvertToMapOfStrings, ReadEnvFile (now
  strings.Cut, which fixes a panic on '='-less lines and truncation of
  values containing '=')
- internal/tmpl: RenderString

Dissolved single-consumer helpers into their consumer:
- Binary type -> runner
- IsExitError -> scheduler (unexported)
- LastLine -> task (unexported)
- MapKeys -> inlined at cmd call sites as slices.Sorted/Collect(maps.Keys),
  removing the reflect-based implementation
- IsURL -> internal/config loader (unexported)

The utils package is deleted. Also makes `list` subcommand output
deterministically sorted.
