+++ cmd/init.go
@@ -14,7 +14,7 @@ import (
 	"github.com/logrusorgru/aurora"
 	"github.com/manifoldco/promptui"
 
-	"github.com/taskctl/taskctl/utils"
+	"github.com/taskctl/taskctl/internal/fsutil"
 )
 
 var configTmpl = `# This is an example of taskctl tasks configuration file.
@@ -73,7 +73,7 @@ func newInitCommand() *cli.Command {
 
 			file := filepath.Join(dir, filename)
 
-			if utils.FileExists(file) {
+			if fsutil.FileExists(file) {
 				replaceConfirmation := promptui.Prompt{
 					Label:     "File already exists. Overwrite",
 					IsConfirm: true,
