+++ internal/config/loader.go
@@ -20,11 +20,21 @@ import (
 	"github.com/pelletier/go-toml"
 	"gopkg.in/yaml.v3"
 
-	"github.com/taskctl/taskctl/utils"
+	"github.com/taskctl/taskctl/internal/fsutil"
 )
 
 // ErrConfigNotFound occurs when requested config file does not exists
 var ErrConfigNotFound = errors.New("config file not found")
+
+// isURL checks if given string is a valid URL
+func isURL(s string) bool {
+	u, err := url.Parse(s)
+	if err != nil {
+		return false
+	}
+
+	return strings.HasPrefix(u.Scheme, "http")
+}
 
 // Loader reads and parses config files
 type Loader struct {
@@ -39,8 +49,8 @@ func NewConfigLoader(dst *Config) Loader {
 	return Loader{
 		dst:     dst,
 		imports: make(map[string]bool),
-		homeDir: utils.MustGetUserHomeDir(),
-		dir:     utils.MustGetwd(),
+		homeDir: fsutil.MustGetUserHomeDir(),
+		dir:     fsutil.MustGetwd(),
 	}
 }
 
@@ -67,7 +77,7 @@ func (cl *Loader) Load(file string) (*Config, error) {
 		}
 	}
 
-	if !utils.IsURL(file) && !filepath.IsAbs(file) {
+	if !isURL(file) && !filepath.IsAbs(file) {
 		file = path.Join(cl.dir, file)
 	}
 
@@ -103,7 +113,7 @@ func (cl *Loader) LoadGlobalConfig() (*Config, error) {
 	}
 
 	file := path.Join(cl.homeDir, ".taskctl", "config.yaml")
-	if !utils.FileExists(file) {
+	if !fsutil.FileExists(file) {
 		return cl.dst, nil
 	}
 
@@ -137,10 +147,10 @@ func (cl *Loader) reset() {
 func (cl *Loader) load(file string) (config map[string]any, err error) {
 	cl.imports[file] = true
 
-	if utils.IsURL(file) {
+	if isURL(file) {
 		config, err = cl.readURL(file)
 	} else {
-		if !utils.FileExists(file) {
+		if !fsutil.FileExists(file) {
 			return config, fmt.Errorf("%s: %w", file, ErrConfigNotFound)
 		}
 		config, err = cl.readFile(file)
@@ -153,7 +163,7 @@ func (cl *Loader) load(file string) (config map[string]any, err error) {
 	importDir := path.Dir(file)
 	if imports, ok := config["import"]; ok && imports != nil {
 		for _, v := range imports.([]any) {
-			if utils.IsURL(v.(string)) {
+			if isURL(v.(string)) {
 				if cl.imports[v.(string)] {
 					continue
 				}
@@ -318,7 +328,7 @@ func (cl *Loader) resolveDefaultConfigFile() (file string, err error) {
 
 		for _, v := range DefaultFileNames {
 			file := filepath.Join(dir, v)
-			if utils.FileExists(file) {
+			if fsutil.FileExists(file) {
 				cl.dir = dir
 				return file, nil
 			}
